+++ src/datagrunt/__init__.py
@@ -28,7 +28,7 @@ __author__ = "Martin Graham"
 __license__ = "MIT"
 
 # Import key classes, functions, or submodules that should be available at the package level
-from datagrunt.csvfiles import CSVReader, CSVWriter, CSVSchemaReportAIGenerated
+from datagrunt.csv_api import CSVReader, CSVWriter, CSVSchemaReportAIGenerated
 
 
 # You can define __all__ to specify what gets imported with "from package import *"
