+++ src/datagrunt/csv_api/__init__.py
@@ -1,7 +1,7 @@
 # Import key classes, functions, or submodules that should be available at the package level
-from datagrunt.csvfiles.csvreader import CSVReader
-from datagrunt.csvfiles.csvwriter import CSVWriter
-from datagrunt.csvfiles.csvai import CSVSchemaReportAIGenerated
+from datagrunt.csv_api.csvreader import CSVReader
+from datagrunt.csv_api.csvwriter import CSVWriter
+from datagrunt.csv_api.csvai import CSVSchemaReportAIGenerated
 
 
 # You can define __all__ to specify what gets imported with "from package import *"
