+++ tests/core_tests/file_io_tests/test_fileproperties.py
